Add dtslint
